+++ src/main/java/com/example/ajoutayo/AjouTayoApplication.java
@@ -2,8 +2,10 @@ package com.example.ajoutayo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class AjouTayoApplication {
 
 	public static void main(String[] args) {
