+++ src/main/java/com/example/ajoutayo/dto/BoardDto.java
@@ -1,0 +1,39 @@
+package com.example.ajoutayo.dto;
+
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+@Entity
+public class BoardDto {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String title;
+
+    private String content;
+    private int count;
+    private Long userId;
+    //@CreatedDate
+    @CreationTimestamp
+    @Column(name = "created_at",updatable = false)
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+}
+
