+++ src/main/java/com/example/ajoutayo/mapper/BoardMapper.java
@@ -1,7 +1,7 @@
-package com.example.ajoutayo.model.dao;
+package com.example.ajoutayo.mapper;
 
 import java.util.List;
-import com.example.ajoutayo.model.dto.BoardDto;
+import com.example.ajoutayo.dto.BoardDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
@@ -13,8 +13,8 @@ public interface BoardMapper {
     int boardCount(); // 곧 생성할 BoardMapper.xml 첫번째 sql 문의 id 와 같음.
     BoardDto getBoard(Long boardId);
 
-    void uploadBoard(BoardDto board);
-    void updateBoard(BoardDto board);
+    void uploadBoard(BoardDto boardDto);
+    void updateBoard(BoardDto boardDto);
     void deleteBoard(Long boardId);
     void viewCount(Long boardId);
