+++ src/main/resources/application.properties
@@ -9,5 +9,5 @@ spring.mvc.pathmatch.matching-strategy=ant_path_matcher
 
 # mybatis
 mybatis.config=mybatis-config.xml
-mybatis.type-aliases-package= com.example.ajoutayo.model
+mybatis.type-aliases-package= com.example.ajoutayo.dto, com.example.ajoutayo.mapper
 mybatis.mapper-locations= mybatis/mapper/*.xml
